src/requirements.txt
aio-pika==6.6.1
aiocli==1.1.1
aioddd==1.0.3
aiohttp==4.0.0a1
bcrypt==3.1.7
cachetools==4.1.1
marshmallow_jsonapi==0.23.2
motor==2.2.0
pyaioapp==1.0.1
pyjwt==1.7.1
uuid==1.30
validators==0.17.1
webargs==6.1.0
